requirements.txt
GitPython==3.1.43
tree_sitter==0.22.3